migrate sql query to stored procedure
